requirements.txt
altair==5.1.2
asgiref==3.7.2
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
clip==0.2.0
cmake==3.27.4.1
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
debugpy==1.7.0
decorator==5.1.1
Django==4.2.5
executing==1.2.0
faiss-cpu==1.7.4
filelock==3.12.3
fonttools==4.42.1
fsspec==2023.9.0
gitdb==4.0.11
GitPython==3.1.40
huggingface-hub==0.17.1
idna==3.4
image==1.5.33
imageio==2.31.6
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipykernel==6.25.2
ipython==8.12.2
japanese-clip @ git+https://github.com/rinnakk/japanese-clip.git@374e08d48b9ba72fcaf7459af4f8c93caffd6fb3
japanize-matplotlib==1.1.3
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.5
lit==16.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl-fft==1.3.6
#mkl-random @ file:///work/mkl/mkl_random_1682950433854/work
mkl-service==2.4.0
mpmath==1.3.0
nest-asyncio==1.5.7
networkx==3.1
nltk==3.8.1
#numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.94
six==1.16.0
smmap==5.0.1
sqlparse==0.4.4
stack-data==0.6.2
streamlit==1.28.1
sympy==1.12
tenacity==8.2.3
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transformers==4.33.1
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.4
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.6
zipp==3.16.2
